refactor: udpate file structure
