+++ src/page/Editor/components/QueryEditor/QueryList/style.tsx
@@ -1,5 +1,4 @@
 import { css } from "@emotion/react"
-import { SerializedStyles } from "@emotion/serialize"
 import { globalColor, illaPrefix } from "@illa-design/theme"
 
 export const QueryListContainer = css`
@@ -7,7 +6,7 @@ export const QueryListContainer = css`
   flex-direction: column;
   width: 255px;
   border-right: 1px solid ${globalColor(`--${illaPrefix}-grayblue-08`)};
-`
+  `
 
 export const QueryListHeader = css`
   height: 48px;
@@ -16,7 +15,7 @@ export const QueryListHeader = css`
   padding: 13px 16px;
   color: ${globalColor(`--${illaPrefix}-grayblue-06`)};
   justify-content: space-between;
-`
+  `
 
 export const NewButton = css`
   background-color: ${globalColor(`--${illaPrefix}-purple-07`)};
@@ -25,7 +24,7 @@ export const NewButton = css`
   font-size: 14px;
   margin: 0 16px;
   height: 32px;
-`
+  `
 
 export const AddIconInNewButton = css`
   margin-right: 8px;
