+++ src/page/Editor/components/QueryEditor/style.tsx
@@ -1,7 +1,4 @@
-export * from "./QueryList"
-export * from "./QueryEditorPanel"
 import { css } from "@emotion/react"
-import { SerializedStyles } from "@emotion/serialize"
 import { globalColor, illaPrefix } from "@illa-design/theme"
 
 export const QueryEditorContainer = css`
